[debug] ros2 only support int_64 arrays for parameters

+++ nav2_controller/include/nav2_controller/plugins/subgoal_included_goal_checker.hpp
@@ -61,7 +61,7 @@ protected:
   };
   
   std::vector<Subgoal> subgoals_;
-  std::vector<int> active_subgoals_;  // Indices of active subgoals
+  std::vector<int64_t> active_subgoals_;  // Indices of active subgoals
   std::mutex subgoals_mutex_;
   bool all_subgoals_reached_;
   double subgoal_tolerance_;
